refactor: idiomatic clippy pedantic fixes
- event_processor.rs: remove redundant else, add separators to long literals
- parser.rs: use u64::from() instead of truncating as-cast
- handlers/dyndns.rs: map().unwrap_or() → map_or()
- handlers/health.rs: map().unwrap_or() → map_or()
- handlers/acme_dns.rs: map().unwrap_or_else() → map_or_else()
- handlers/acme_order.rs: same pattern

+++ nodns-bot-rs/src/parser.rs
@@ -61,10 +61,10 @@ pub fn classify_event(
     block_private_ip: bool,
     max_txt_length: usize,
 ) -> Result<ParsedEvent, ParserError> {
-    if event.kind.as_u16() != KIND_DNS_RECORD as u16 {
+    if u64::from(event.kind.as_u16()) != KIND_DNS_RECORD {
         return Err(ParserError::WrongKind {
             expected: KIND_DNS_RECORD,
-            got: event.kind.as_u16() as u64,
+            got: u64::from(event.kind.as_u16()),
         });
     }
